Fix DateTimeTrigger not triggering on restored state (#3277)
During startup most likely the item state is `UndefType.NULL` which results in no triggering. States are restored from persistence by directly setting the item state, not by an `ItemStateEvent` and were therefore missed. Since changing the trigger is not necessary when the state updates but does not change, the event listening was changed to `ItemStateChangedEvent` only.

Signed-off-by: Jan N. Klug <github@klug.nrw>