[REST] Fix API to get file format when applied to a subset of elements (#4786)
Fix #4778

Things and items are now sorted in output only when all elements are requested.
If the API is requested only for a subset of elements (things or items), the output respects the input order of elements.

DSL generator for things and items is also fixed to avoid a crash when called with 0 element.

Signed-off-by: Laurent Garnier <lg.hc@free.fr>