YAML config: fix user defined channels creation when binding loads late
While things can now be loaded even when the binding is not yet available, channel types for user defined channels could be not yet in the registry when these channels are first created.
So when the binding is finally available, the thing is updated and user defined channels are now created again to retrieve properties from channel types.

This fix avoids warnings at startup regarding missing channel types.

Signed-off-by: Laurent Garnier <lg.hc@free.fr>