Isolate discovery result registrations from binding threads (#5032)
By using the discovery thread pool to do the actual registrations in AbstractDiscoveryService, binding threads won't have to wait for the registration to complete, which can be slow.

In addition, some thread-safety fixes have been done in AbstractDiscoveryService and DiscoveryServiceRegistryImpl, both to avoid contention and to ensure isolation of mutable objects.

Signed-off-by: Ravi Nadahar <nadahar@rediffmail.com>